Move EndpointProtocol/SocketType to Core
